share notArchivedFilter in more places
