+++ apps/hash-frontend/src/shared/notifications-context.tsx
@@ -8,6 +8,7 @@ import {
 import {
   currentTimeInstantTemporalAxes,
   generateVersionedUrlMatchingFilter,
+  notArchivedFilter,
   zeroedGraphResolveDepths,
 } from "@local/hash-isomorphic-utils/graph-queries";
 import { systemTypes } from "@local/hash-isomorphic-utils/ontology-types";
@@ -137,38 +138,7 @@ export const NotificationsContextProvider: FunctionComponent<
           systemTypes.entityType.notification.entityTypeId,
           { ignoreParents: false },
         ),
-        {
-          any: [
-            {
-              equal: [
-                {
-                  path: [
-                    "properties",
-                    extractBaseUrl(
-                      systemTypes.propertyType.archived.propertyTypeId,
-                    ),
-                  ],
-                },
-                // @ts-expect-error -- We need to update the type definition of `EntityStructuralQuery` to allow for this
-                //   @see https://linear.app/hash/issue/H-1207
-                null,
-              ],
-            },
-            {
-              equal: [
-                {
-                  path: [
-                    "properties",
-                    extractBaseUrl(
-                      systemTypes.propertyType.archived.propertyTypeId,
-                    ),
-                  ],
-                },
-                { parameter: false },
-              ],
-            },
-          ],
-        },
+        notArchivedFilter,
       ],
     }),
     [authenticatedUser],
